+++ allele_components/separators.txt
@@ -1,0 +1,5 @@
+--
+::
+<<
+-
+:
